add SitePermissions as inline table to a users haztrak profile, remove admin_rcrainfo_profile from HaztrakProfile
